requirements.txt
django-environ==0.6.0
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
djangorestframework-jwt==1.11.0
